+++ src/main/java/kusitms/jangkku/domain/auth/dto/NaverUserInfo.java
@@ -1,23 +1,23 @@
-package kusitms.jangkku.global.auth.dto;
+package kusitms.jangkku.domain.auth.dto;
 
-import kusitms.jangkku.global.common.constant.Provider;
+import kusitms.jangkku.domain.auth.constant.Provider;
 import lombok.AllArgsConstructor;
 
 import java.util.Map;
 
 @AllArgsConstructor
-public class GoogleUserInfo implements OAuth2UserInfo {
+public class NaverUserInfo implements OAuth2UserInfo {
 
     private Map<String, Object> attributes;
 
     @Override
     public String getProviderId() {
-        return (String) attributes.get("sub");
+        return (String) attributes.get("id");
     }
 
     @Override
     public String getProvider() {
-        return Provider.GOOGLE_PROVIDER.getProvider();
+        return Provider.NAVER_PROVIDER.getProvider();
     }
 
     @Override
