+++ src/main/java/kusitms/jangkku/domain/user/api/UserController.java
@@ -2,6 +2,7 @@ package kusitms.jangkku.domain.user.api;
 
 import jakarta.servlet.http.HttpServletResponse;
 import kusitms.jangkku.domain.user.application.UserService;
+import kusitms.jangkku.domain.user.constant.UserSuccessStatus;
 import kusitms.jangkku.domain.user.dto.UserDto;
 import kusitms.jangkku.domain.user.exception.UserErrorResult;
 import kusitms.jangkku.global.common.ApiResponse;
@@ -18,13 +19,13 @@ public class UserController {
 
     // 유저 기본 정보를 입력 받아 저장하는 API
     @PostMapping("/register")
-    public ResponseEntity<ApiResponse<UserDto.UserRegisterResponse>> RegisterUser(
+    public ResponseEntity<ApiResponse<UserDto.UserRegisterResponse>> registerUser(
             HttpServletResponse response,
             @RequestHeader("Authorization") String authorizationHeader,
             @RequestBody UserDto.UserRegisterRequest userRegisterRequest) {
 
         UserDto.UserRegisterResponse userRegisterResponse = userService.registerUser(response, authorizationHeader, userRegisterRequest);
-        return ApiResponse.onSuccess(SuccessStatus.SUCCESS_REGISTER_USER, userRegisterResponse);
+        return ApiResponse.onSuccess(UserSuccessStatus.SUCCESS_REGISTER_USER, userRegisterResponse);
     }
 
     // 닉네임 중복 여부를 반환하는 API
@@ -33,7 +34,7 @@ public class UserController {
             @PathVariable("nickname") String nickname) {
 
         if (!userService.isDuplicate(nickname)) {
-            return ApiResponse.onSuccess(SuccessStatus.IS_NICKNAME_POSSIBLE);
+            return ApiResponse.onSuccess(UserSuccessStatus.IS_NICKNAME_POSSIBLE);
         } else {
             return ApiResponse.onFailure(UserErrorResult.IS_DUPLICATE_NICKNAME);
         }
