+++ src/main/java/kusitms/jangkku/global/common/constant/SuccessStatus.java
@@ -11,12 +11,7 @@ import org.springframework.http.HttpStatus;
 public enum SuccessStatus implements BaseCode {
     // 전역 응답 코드
     _OK(HttpStatus.OK, "200", "성공입니다."),
-    _CREATED(HttpStatus.CREATED, "201", "생성에 성공했습니다."),
-
-    // 유저 & 토큰
-    SUCCESS_REGISTER_USER(HttpStatus.CREATED, "201", "유저 등록에 성공했습니다."),
-    IS_NICKNAME_POSSIBLE(HttpStatus.OK, "200", "사용할 수 있는 닉네임입니다."),
-    CREATED_ACCESS_TOKEN(HttpStatus.CREATED, "201", "액세스 토큰 재발행에 성공했습니다.");
+    _CREATED(HttpStatus.CREATED, "201", "생성에 성공했습니다.");
 
     private final HttpStatus httpStatus;
     private final String code;
